+++ data/result-exporter/swe-cne-exporter/src/test/java/com/powsybl/openrao/data/swecneexporter/SweCneTest.java
@@ -8,18 +8,15 @@
 package com.powsybl.openrao.data.swecneexporter;
 
 import com.powsybl.iidm.network.Network;
-import com.powsybl.openrao.data.cneexportercommons.CneExporterParameters;
 import com.powsybl.openrao.data.cracapi.Crac;
 import com.powsybl.openrao.data.cracapi.CracCreationContext;
 import com.powsybl.openrao.data.cracapi.parameters.CracCreationParameters;
-import com.powsybl.openrao.data.cracio.cim.craccreator.CimCracCreationContext;
 import com.powsybl.openrao.data.cracio.cim.parameters.CimCracCreationParameters;
 import com.powsybl.openrao.data.cracio.cim.parameters.RangeActionSpeed;
 import com.powsybl.openrao.data.raoresultapi.RaoResult;
 import com.powsybl.openrao.monitoring.anglemonitoring.AngleMonitoringResult;
 import com.powsybl.openrao.monitoring.anglemonitoring.RaoResultWithAngleMonitoring;
 import com.powsybl.openrao.monitoring.anglemonitoring.json.AngleMonitoringResultImporter;
-import com.powsybl.openrao.raoapi.parameters.RaoParameters;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -33,6 +30,7 @@ import org.xmlunit.diff.Difference;
 import java.io.*;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
+import java.util.Properties;
 import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -47,6 +45,7 @@ class SweCneTest {
     private Network network;
     private RaoResultWithAngleMonitoring raoResultWithAngle;
     private RaoResultWithAngleMonitoring raoResultFailureWithAngle;
+    private Properties properties;
 
     @BeforeEach
     public void setUp() throws IOException {
@@ -85,17 +84,23 @@ class SweCneTest {
         RaoResult raoResultWithFailure = RaoResult.read(inputStream3, crac);
         raoResultWithAngle = new RaoResultWithAngleMonitoring(raoResult, angleMonitoringResult);
         raoResultFailureWithAngle = new RaoResultWithAngleMonitoring(raoResultWithFailure, angleMonitoringResult);
+
+        properties = new Properties();
+        properties.setProperty("document-id", "documentId");
+        properties.setProperty("revision-number", "1");
+        properties.setProperty("domain-id", "domainId");
+        properties.setProperty("process-type", "Z01");
+        properties.setProperty("sender-id", "senderId");
+        properties.setProperty("sender-role", "A04");
+        properties.setProperty("receiver-id", "receiverId");
+        properties.setProperty("receiver-role", "A36");
+        properties.setProperty("time-interval", "2021-04-02T12:00:00Z/2021-04-02T13:00:00Z");
     }
 
     @Test
     void testExport() {
-        CneExporterParameters params = new CneExporterParameters(
-            "documentId", 1, null, CneExporterParameters.ProcessType.Z01,
-            "senderId", CneExporterParameters.RoleType.SYSTEM_OPERATOR,
-            "receiverId", CneExporterParameters.RoleType.CAPACITY_COORDINATOR,
-            "2021-04-02T12:00:00Z/2021-04-02T13:00:00Z");
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        new SweCneExporter().exportCne(crac, (CimCracCreationContext) cracCreationContext, raoResultWithAngle, new RaoParameters(), params, outputStream);
+        new SweCneExporter().exportData(raoResultWithAngle, cracCreationContext, properties, outputStream);
         try {
             InputStream inputStream = new FileInputStream(SweCneTest.class.getResource("/SweCNE_Z01.xml").getFile());
             compareCneFiles(inputStream, new ByteArrayInputStream(outputStream.toByteArray()));
@@ -106,13 +111,8 @@ class SweCneTest {
 
     @Test
     void testExportWithFailure() {
-        CneExporterParameters params = new CneExporterParameters(
-                "documentId", 1, null, CneExporterParameters.ProcessType.Z01,
-                "senderId", CneExporterParameters.RoleType.SYSTEM_OPERATOR,
-                "receiverId", CneExporterParameters.RoleType.CAPACITY_COORDINATOR,
-                "2021-04-02T12:00:00Z/2021-04-02T13:00:00Z");
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        new SweCneExporter().exportCne(crac, (CimCracCreationContext) cracCreationContext, raoResultFailureWithAngle, new RaoParameters(), params, outputStream);
+        new SweCneExporter().exportData(raoResultFailureWithAngle, cracCreationContext, properties, outputStream);
         try {
             InputStream inputStream = new FileInputStream(SweCneTest.class.getResource("/SweCNEWithFailure_Z01.xml").getFile());
             compareCneFiles(inputStream, new ByteArrayInputStream(outputStream.toByteArray()));
